Refactored screens to /screens folder
