+++ app/src/main/java/com/example/bluetooth_chat/ui/screens/BluetoothDevicesScreen.kt
@@ -1,4 +1,4 @@
-package com.example.bluetooth_chat.ui.components
+package com.example.bluetooth_chat.ui.screens
 
 import android.Manifest
 import android.bluetooth.BluetoothDevice
