+++ app/src/main/java/com/example/bluetooth_chat/ui/screens/GroupsScreen.kt
@@ -1,4 +1,4 @@
-package com.example.bluetooth_chat.ui.components
+package com.example.bluetooth_chat.ui.screens
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.Text
